Avoid premature case modification
Case should be preserved. Parameter order and duplication is preserved, and a
map is added to content-type and media-type for convenience. The map keys are
transformed to lower-case to support case-insensitivity in lookups.

The intention here is that reap should preserve information as far as possible
in order to support re-encoding. The responsibility for case-insensitivity is
deferred to the caller.