Breaking change: Change namespace names
After a lot of consideration I feel that the previous juxt.http namespace is too
broad. The resulting structures of parsing strings should be governed in the
juxt.reap.alpha namespace.

The underlying issue is that desire to allow performance-critical code to use
reap structures so that libraries can distinguish between data that may need
parsing with data that has already been parsed.

The example that is driving this change is the problem of whether to call the
pick library's API with strings (which would require parsing by reap for pick to
act upon them) or with structures that are the result of parsing with reap. The
former is certainly much more convenient and less error prone, while the latter
allows users to optimise performance. The resolution of this problem is to allow
'both'. In the pick library, a higher-level API can be used which can be given
strings (e.g. media-types such as "text/html;charset=utf-8") and a lower-level
API which requires the structures resulting from parsing with reap.