Ensure matcher is reset on each alternative
A java.util.regex.Matcher is mutable, which means that matching in one
branch (of alternatives) affects the outcome of other branches. This is
obviously not desirable and introduces bugs.

This bug was detected in the processing GraphQL wrapping types because the
bang (!) is consumed after the type it is wrapping.