Default to using keywords but allow strings
This is a controversial design choice, since client code shouldn't be
inconsistent in its choice to use strings or keywords.

For various justifiable design reasons, some programs may elect to use
string keys when using URI Templates. We shouldn't be overly strict
and force callers to have to convert to keywords.

We may revisit this design choice in future. To be future-proof, use
keywords consistently.